requirements.txt
# Dev environment
pip
autopep8

# App
flask

python-dotenv
requests
google-generativeai
pymongo